fix(nextra-theme-docs): moved details and summery out of mdx components and extended them
