+++ examples/nextra/pages/docs/advanced.mdx
@@ -1,0 +1,29 @@
+# Advanced features!
+
+You can create this index page now!
+
+## Checkboxes
+
+- [x] First <strong>thing</strong>
+- [x] Second _thing_
+
+## Details and Summary
+
+<details>
+  <summary>**Version History**</summary>
+  | Version   | Changes                                                             |
+  | --------- | ------------------------------------------------------------------- |
+  | `v10.0.0` | `locale`, `locales`, `defaultLocale`, and `notFound` options added. |
+  | `v9.3.0`  | `getServerSideProps` introduced.                                    |
+  <details open>
+    <summary>**Second Level Details**</summary>
+    Hello!
+  </details>
+  <details>
+    <summary>Second _Level_ Details</summary>
+    Hello!
+  </details>
+  [google](https://google.com)
+</details>
+
+hey!!
